refactor: moved shared types into backend app
