+++ apps/backend/src/api/v1/types/Favorite.ts
