--- shared/types/SharedState.ts
@@ -1,4 +1,0 @@
-export type SharedState = {
-  message: string;
-  visits: number;
-};
